flink/requirements.txt
python-dotenv
apache-flink==1.17.2